+++ lib/src/models/artist_page_details.g.dart
@@ -6,6 +6,23 @@ part of 'artist_page_details.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
+ArtistUrls _$ArtistUrlsFromJson(Map<String, dynamic> json) => ArtistUrls(
+  albums: json['albums'] as String?,
+  bio: json['bio'] as String?,
+  comments: json['comments'] as String?,
+  songs: json['songs'] as String?,
+  overview: json['overview'] as String?,
+);
+
+Map<String, dynamic> _$ArtistUrlsToJson(ArtistUrls instance) =>
+    <String, dynamic>{
+      'albums': instance.albums,
+      'bio': instance.bio,
+      'comments': instance.comments,
+      'songs': instance.songs,
+      'overview': instance.overview,
+    };
+
 ArtistPageDetails _$ArtistPageDetailsFromJson(Map<String, dynamic> json) =>
     ArtistPageDetails(
       artistId: json['artistId'] as String,
@@ -28,6 +45,18 @@ ArtistPageDetails _$ArtistPageDetailsFromJson(Map<String, dynamic> json) =>
       isRadioPresent: json['isRadioPresent'] as bool?,
       isFollowed: json['is_followed'] as bool?,
       fanCount: json['fan_count'] as String?,
+      dob: json['dob'] as String?,
+      fb: json['fb'] as String?,
+      twitter: json['twitter'] as String?,
+      wiki: json['wiki'] as String?,
+      urls: json['urls'] == null
+          ? null
+          : ArtistUrls.fromJson(json['urls'] as Map<String, dynamic>),
+      availableLanguages: (json['availableLanguages'] as List<dynamic>?)
+          ?.map((e) => e as String)
+          .toList(),
+      topEpisodes: json['topEpisodes'] as List<dynamic>?,
+      modules: json['modules'] as Map<String, dynamic>?,
     );
 
 Map<String, dynamic> _$ArtistPageDetailsToJson(ArtistPageDetails instance) =>
@@ -52,4 +81,12 @@ Map<String, dynamic> _$ArtistPageDetailsToJson(ArtistPageDetails instance) =>
       'isRadioPresent': instance.isRadioPresent,
       'is_followed': instance.isFollowed,
       'fan_count': instance.fanCount,
+      'dob': instance.dob,
+      'fb': instance.fb,
+      'twitter': instance.twitter,
+      'wiki': instance.wiki,
+      'urls': instance.urls,
+      'availableLanguages': instance.availableLanguages,
+      'topEpisodes': instance.topEpisodes,
+      'modules': instance.modules,
     };